gene set resource: update tests
