+++ dae/dae/gene_sets/tests/test_gene_sets_db.py
@@ -1,153 +1,12 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
-import textwrap
-from pathlib import Path
-from typing import Any
-
-import pytest
-
 from dae.gene_sets.gene_sets_db import (
     GeneSetCollection,
     GeneSetsDb,
     build_gene_set_collection_from_resource_id,
 )
 from dae.genomic_resources.repository import (
-    GR_CONF_FILE_NAME,
     GenomicResourceRepo,
 )
-from dae.genomic_resources.testing import (
-    build_filesystem_test_repository,
-    build_inmemory_test_repository,
-    convert_to_tab_separated,
-    setup_directories,
-)
-
-
-@pytest.fixture
-def gene_sets_repo_fixture(
-    grr_contents,
-    tmp_path_factory: pytest.TempPathFactory,
-) -> tuple[Path, GenomicResourceRepo]:
-    root_path = tmp_path_factory.mktemp("gene_sets_repo_tests")
-    setup_directories(root_path, grr_contents)
-    return root_path, build_filesystem_test_repository(root_path)
-
-
-@pytest.fixture
-def gene_sets_repo_path(  # noqa: FURB118
-    gene_sets_repo_fixture,
-) -> Path:
-    return gene_sets_repo_fixture[0]
-
-
-@pytest.fixture
-def gene_sets_repo_in_memory(grr_contents) -> GenomicResourceRepo:
-    return build_inmemory_test_repository(grr_contents)
-
-
-@pytest.fixture
-def grr_contents() -> dict[str, Any]:
-    return {
-        "main": {
-            GR_CONF_FILE_NAME: textwrap.dedent("""
-                type: gene_set
-                id: main
-                format: directory
-                directory: GeneSets
-                web_label: Main
-                web_format_str: "key| (|count|): |desc"
-                """),
-            "GeneSets": {
-                "main_candidates.txt": (
-                    "main_candidates\n"
-                    "Main Candidates\n"
-                    "POGZ\n"
-                    "CHD8\n"
-                    "ANK2\n"
-                    "FAT4\n"
-                    "NBEA\n"
-                    "CELSR1\n"
-                    "USP7\n"
-                    "GOLGA5\n"
-                    "PCSK2\n"
-                ),
-                "alt_candidates.txt": (
-                    "alt_candidates\n"
-                    "Alt Candidates\n"
-                    "DIABLO\n"
-                ),
-            },
-        },
-        "test_mapping": {
-            GR_CONF_FILE_NAME: textwrap.dedent("""
-                type: gene_set
-                id: test_mapping
-                format: map
-                filename: test-map.txt
-                web_label: Test mapping
-                web_format_str: "key| (|count|)"
-            """),
-            "test-map.txt": (
-                "#geneNS\tsym\n"
-                "POGZ\ttest:01 test:02\n"
-                "CHD8\ttest:02 test:03\n"
-            ),
-            "test-mapnames.txt": (
-                "test:01\ttest_first\n"
-                "test:02\ttest_second\n"
-                "test:03\ttest_third\n"
-            ),
-        },
-        "test_gmt": {
-            GR_CONF_FILE_NAME: textwrap.dedent("""
-                type: gene_set
-                id: test_gmt
-                format: gmt
-                filename: test-gmt.gmt
-                web_label: Test GMT
-                web_format_str: "key| (|count|)"
-            """),
-            "test-gmt.gmt": (
-                "TEST_GENE_SET1\tsomedescription\tPOGZ\tCHD8\n"
-                "TEST_GENE_SET2\tsomedescription\tANK2\tFAT4\n"
-                "TEST_GENE_SET3\tsomedescription\tPOGZ\n"
-            ),
-        },
-        "test": {
-            GR_CONF_FILE_NAME: textwrap.dedent("""
-                type: gene_set
-                id: test_mapping
-                format: map
-                filename: test-map.txt
-                web_label: Test mapping
-                web_format_str: "key| (|count|)"
-            """),
-            "test-map.txt": convert_to_tab_separated("""
-                #geneNS tsym
-                POGZ    test:01||test:02
-                CHD8    test:02||test:03
-            """),
-            "test-mapnames.txt": convert_to_tab_separated("""
-                test:01  test_first
-                test:02  test_second
-                test:03  test_third
-            """),
-        },
-    }
-
-
-@pytest.fixture
-def gene_sets_db(
-    gene_sets_repo_in_memory: GenomicResourceRepo,
-) -> GeneSetsDb:
-    resources = [
-        gene_sets_repo_in_memory.get_resource("main"),
-        gene_sets_repo_in_memory.get_resource("test_mapping"),
-        gene_sets_repo_in_memory.get_resource("test_gmt"),
-    ]
-    gene_set_collections = [
-        GeneSetCollection(r) for r in resources
-    ]
-    return GeneSetsDb(gene_set_collections)
 
 
 def test_gene_set_collection_main(
