+++ dae/dae/gene_sets/tests/test_gene_sets_impl.py
@@ -1,17 +1,37 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
 from pathlib import Path
 
-import pytest_mock
+import pytest
 
 from dae.gene_sets.gene_sets_db import (
     build_gene_set_collection_from_resource_id,
 )
 from dae.gene_sets.implementations.gene_sets_impl import GeneSetCollectionImpl
-from dae.genomic_resources.cli import cli_manage
 from dae.genomic_resources.repository import (
     GenomicResourceRepo,
 )
+from dae.genomic_resources.testing import (
+    build_filesystem_test_repository,
+    setup_directories,
+)
 from dae.task_graph.graph import TaskGraph
+
+
+@pytest.fixture
+def gene_sets_repo_fixture(
+    grr_contents,
+    tmp_path_factory: pytest.TempPathFactory,
+) -> tuple[Path, GenomicResourceRepo]:
+    root_path = tmp_path_factory.mktemp("gene_sets_repo_tests")
+    setup_directories(root_path, grr_contents)
+    return root_path, build_filesystem_test_repository(root_path)
+
+
+@pytest.fixture
+def gene_sets_repo_path(  # noqa: FURB118
+    gene_sets_repo_fixture,
+) -> Path:
+    return gene_sets_repo_fixture[0]
 
 
 def test_add_statistics_build_tasks(
@@ -29,28 +49,4 @@ def test_add_statistics_build_tasks(
     assert len(graph.tasks) == 0
 
     gene_sets_collection_impl.add_statistics_build_tasks(graph)
-    assert len(graph.tasks) == 0
-
-
-def test_calc_statistics_hash(
-    gene_sets_repo_path: Path,
-    mocker: pytest_mock.MockerFixture,
-) -> None:
-    calc_statistics_hash_mock = mocker.patch(
-        "dae.gene_sets.gene_sets_db.GeneSetCollectionImpl.calc_statistics_hash",
-        autospec=True,
-    )
-    calc_statistics_hash_mock.return_value = b""
-
-    add_statistics_build_tasks_mock = mocker.patch(
-        "dae.gene_sets.gene_sets_db.GeneSetCollectionImpl.add_statistics_build_tasks",
-        autospec=True,
-    )
-    add_statistics_build_tasks_mock.return_value = b""
-
-    cli_manage([
-        "repo-repair", "-R", str(gene_sets_repo_path), "-j", "1",
-    ])
-
-    assert calc_statistics_hash_mock.called is True
-    assert add_statistics_build_tasks_mock.called is True
+    assert len(graph.tasks) == 1
